ordering, dashboard, etc

+++ ocnds-laravel/database/migrations/2024_08_19_143246_create_phosphorylations_table.php
@@ -21,10 +21,10 @@ return new class extends Migration
             $table->float('k198r_m')->nullable();
             $table->float('r47g')->nullable();
             $table->float('d156e')->nullable();
-            $table->float('pval_k198rs')->nullable();
-            $table->float('pval_k198rm')->nullable();
-            $table->float('pval_r47g')->nullable();
-            $table->float('pval_d156e')->nullable();
+            $table->double('pval_k198rs')->nullable();
+            $table->double('pval_k198rm')->nullable();
+            $table->double('pval_r47g')->nullable();
+            $table->double('pval_d156e')->nullable();
             $table->string('modified_residue')->nullable();
             $table->integer('modified_position')->nullable();
             $table->string('15mer')->nullable();
